Refactor: Extract Farmhand module helpers into FarmhandService
