+++ src/components/Farmhand/FarmhandService.test.ts
@@ -1,29 +1,31 @@
 import { sprinkler } from '../../data/items.js'
 import { carrotSeed } from '../../data/crops/carrot.js'
 
-import {
-  computePlayerInventory,
-  getFieldToolInventory,
-  getPlantableCropInventory,
-} from './Farmhand.js'
+import { FarmhandService } from './FarmhandService.js'
 
 vitest.mock('../../data/maps.js')
 vitest.mock('../../data/items.js')
 
-describe('private helpers', () => {
+describe('FarmhandService', () => {
   describe('computePlayerInventory', () => {
     const inventory = [{ quantity: 1, id: 'carrot-seed' }]
 
     test('computes inventory with no value adjustments', () => {
-      const playerInventory = computePlayerInventory(inventory, {})
+      const playerInventory = FarmhandService.computePlayerInventory(
+        inventory,
+        {}
+      )
 
       expect(playerInventory).toEqual([{ quantity: 1, ...carrotSeed }])
     })
 
     test('computes inventory with value adjustments', () => {
-      const playerInventory = computePlayerInventory(inventory, {
-        'carrot-seed': 2,
-      })
+      const playerInventory = FarmhandService.computePlayerInventory(
+        inventory,
+        {
+          'carrot-seed': 2,
+        }
+      )
 
       expect(playerInventory).toEqual([
         { ...carrotSeed, quantity: 1, value: 30 },
@@ -33,7 +35,7 @@ describe('private helpers', () => {
 
   describe('getFieldToolInventory', () => {
     test('selects field tools from inventory', () => {
-      const fieldToolInventory = getFieldToolInventory([
+      const fieldToolInventory = FarmhandService.getFieldToolInventory([
         { ...sprinkler, quantity: 1 },
         { ...carrotSeed, quantity: 1 },
       ])
@@ -48,7 +50,9 @@ describe('private helpers', () => {
         { id: 'carrot-seed', quantity: 1 },
         { id: 'weed', quantity: 1 },
       ]
-      const plantableCropInventory = getPlantableCropInventory(inventory)
+      const plantableCropInventory = FarmhandService.getPlantableCropInventory(
+        inventory
+      )
 
       expect(plantableCropInventory).toEqual([{ ...carrotSeed, quantity: 1 }])
     })
